Remove ObjectUtils module, rename SchemaUtils to Schema
