+++ .eslintrc
@@ -105,7 +105,7 @@
     "new-parens": 1,                  // Disallow the omission of parentheses when invoking a constructor with no arguments
     "newline-after-var": 0,           // Allow/disallow an empty newline after var statement (off by default)
     "no-array-constructor": 1,        // Disallow use of the Array constructor
-    "no-inline-comments": 1,          // Disallow comments inline after code (off by default)
+    "no-inline-comments": 0,          // Disallow comments inline after code (off by default)
     "no-lonely-if": 1,                // Disallow if as the only statement in an else block (off by default)
     "no-mixed-spaces-and-tabs": 1,    // Disallow mixed spaces and tabs for indentation
     "no-multiple-empty-lines": [1, {"max": 2}], // disallow multiple empty lines (off by default)
@@ -124,7 +124,7 @@
     "semi": [1, "always"],                // Require or disallow use of semicolons instead of ASI
     "semi-spacing": [1, {"before": false, "after": true}], // enforce spacing before and after semicolons
     "space-before-blocks": [1, "always"], // Require or disallow space before blocks (off by default)
-    "space-in-brackets": [1, "never"],    // Require or disallow spaces inside brackets (off by default)
+    "space-in-brackets": [0, "never"],    // Require or disallow spaces inside brackets (off by default)
     "space-in-parens": [1, "never"],      // Require or disallow spaces inside parentheses (off by default)
     "space-infix-ops": [1, {}],     // Require spaces around operators
     "space-return-throw-case": [1], // require a space after return, throw, and case
