+++ src/invariant.js
@@ -2,7 +2,7 @@
  * @copyright 2015, Prometheus Research, LLC
  */
 
-export default function invariant(cond, msg, a, b, c, d, e, f) {
+export default function invariant(cond, msg, a, b, c, d, e, f) { // eslint-disable-line max-params
   if (!cond) {
     if (msg) {
       let replacements = [a, b, c, d, e, f];
